% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_function.R
\name{gps_function}
\alias{gps_function}
\title{The gps_function processes a single CSV file generated by a GPS logger device.
The input file is expected to contain one row per time point (e.g., minute), with each row including GPS coordinates (latitude and longitude), timestamp details (date and time), and capturing the complete set of GPS data for a single participant over the course of the study.}
\usage{
gps_function(data_directory, high.speed = 130, height.change = 1000,
             impute.gap.size = 5, imputation = T,
             counties = c("illinois,cook", "illinois,dekalb",
                          "illinois,du page","illinois,grundy",
                          "illinois,kane", "illinois,kendall",
                          "illinois,lake", "illinois,mchenry",
                          "illinois,will", "indiana,jasper",
                          "indiana,lake", "indiana,newton",
                          "indiana,porter", "wisconsin,kenosha"))
}
\arguments{
\item{data_directory}{a string of working directory (should be a .csv file)}

\item{high.speed}{speed higher than this value (km/h) will be dropped (default: 130)}

\item{height.change}{one-min height change higher than this value (m) will be dropped (default: 1000)}

\item{impute.gap.size}{Size (mins) of the gaps to impute (should be less than 30 mins) (default: 5)}

\item{imputation}{whether or not to save the imputed GPS data file (default: True)}

\item{counties}{a string vector of county names to be investigated (default: "illinois,cook", "illinois,dekalb", "illinois,du page","illinois,grundy",
"illinois,kane", "illinois,kendall", "illinois,lake", "illinois,mchenry",
"illinois,will", "indiana,jasper", "indiana,lake", "indiana,newton",
"indiana,porter", "wisconsin,kenosha")}
}
\value{
The code \code{gps_function} returns a brief summary of the cleaned and imputed GPS data, automatically saving the processed file to the working directory.
}
\description{
The gps_function processes a single CSV file generated by a GPS logger device.
The input file is expected to contain one row per time point (e.g., minute), with each row including GPS coordinates (latitude and longitude), timestamp details (date and time), and capturing the complete set of GPS data for a single participant over the course of the study.
}
