% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_function.R
\name{gt3x_function}
\alias{gt3x_function}
\title{This gt3x_function processes a single GT3X file to generate two output files:
(1) a CSV file with minute-level epochs of activity categories for each minute of the participant's usable data, and
(2) a CSV file summarizing daily wake/sleep times, wear/non-wear times, and inclusion decisions for each study day.}
\usage{
gt3x_function(data_directory, valid_day_criteria = 10,
              wake_bout = 1, sleep_bout = 1,
              non_wear_chunk_min = 45, start_hour = "04:00:00",
              participant_id = "P2E30001", days_include = 21)
}
\arguments{
\item{data_directory}{a string of working directory (should be a gt3x file)}

\item{valid_day_criteria}{minimum hours to be considered as a valid day (default: 10)}

\item{wake_bout}{the number of consecutive minutes with positive activity counts to define wake time (default: 1)}

\item{sleep_bout}{the number of consecutive minutes with positive activity counts to define wake time (default: 1)}

\item{non_wear_chunk_min}{the number of inactive minutes to be considered as non-wear time (default: 45)}

\item{start_hour}{the starting hour of a day (default: "04:00:00")}

\item{participant_id}{a string of participant ID (default: "P2E30001")}

\item{days_include}{the number of days to be included in this study (default: 21)}
}
\value{
The code \code{gt3x_function} generates a csv file with minute-level epochs of activity categories for each minute of the participant's usable data
and a csv file summarizing daily wake/sleep times, wear/non-wear times, and inclusion decisions for each study day.
}
\description{
This gt3x_function processes a single GT3X file to generate two output files:
(1) a CSV file with minute-level epochs of activity categories for each minute of the participant's usable data, and
(2) a CSV file summarizing daily wake/sleep times, wear/non-wear times, and inclusion decisions for each study day.
}
