% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_function.R
\name{harmonize_function}
\alias{harmonize_function}
\title{The harmonize_function merges the output data files from gps_function and gt3x_function by day and time (e.g., per minute),
keeping all rows from both files and displaying missing data for either accelerometer or GPS data as applicable.
This allows for straightforward editing or post-processing to eliminate undesired rows, producing a single CSV file with one row per minute that includes both latitude/longitude and activity classification.}
\usage{
harmonize_function(acti_data, gps_data, participant_id = "P2E30001")
}
\arguments{
\item{acti_data}{a csv file output generated by the gt3x_function with activity classification}

\item{gps_data}{a csv file output generated by the gps_function}

\item{participant_id}{a string of participant ID (default: "P2E30001")}
}
\value{
This function saves a csv file with merged output from gps_function and gt3x_function by time.
}
\description{
The harmonize_function merges the output data files from gps_function and gt3x_function by day and time (e.g., per minute),
keeping all rows from both files and displaying missing data for either accelerometer or GPS data as applicable.
This allows for straightforward editing or post-processing to eliminate undesired rows, producing a single CSV file with one row per minute that includes both latitude/longitude and activity classification.
}
